reverting datatype change to be able to release this change before version 3.0

+++ src/main/java/redis/clients/jedis/BinaryJedis.java
@@ -2866,7 +2866,7 @@ public class BinaryJedis implements BasicCommands, BinaryJedisCommands, MultiKey
     }
   }
 
-  public int getDB() {
+  public Long getDB() {
     return client.getDB();
   }
